Story 9-15: code-review patches (review → done)
3 parallel reviewers (Blind/Edge/Auditor) — 4 actionable patches +
12 deferred + ~10 dismissed. Acceptance Auditor: ALL 14 ACs PASS.
0 BLOCKERS, 0 decision-needed.

P1 (edge, HIGH) — Borrowers empty-state CTA wired to JS toggle.
The `<a href="#add-form">` anchor would scroll the page but leave the
form display:none (hidden by Tailwind `.hidden`) — UX dead-end on the
encouraging empty state. Extended `static/js/borrowers.js` with a
delegated listener on `a[href="#add-form"]` calling the same showForm
handler the existing button uses. Idempotent via dataset.wired guard.

P2 (auditor, HIGH) — `src/routes/home.rs` split to keep under
Foundation Rule #12 ceiling. home.rs was at 2018 LOC post-implementation.
Extracted `render_search_fragment`, `render_search_row`,
`render_pagination_oob`, `render_empty_state` + `SearchEmptyState`
template struct into a NEW sibling module
`src/routes/home_search_fragment.rs` (290 LOC, pub(super)-only). Result:
home.rs now 1780 LOC (-238). The 2 existing test cases moved to the new
module; the 3 NEW empty-state tests live there too.

P3 (edge, MEDIUM) — Live HTMX search empty-state filter sub-case added.
`render_empty_state` previously took only `(query, is_librarian, loc)`
and ALWAYS rendered the search-empty copy. When a user filtered via
HTMX with empty query, they saw "No matches for ''. Try a broader term."
with broken empty quotes + irrelevant CTA — divergent from the navigated
path. Extended signature to `(query, has_filter, is_librarian, loc)`.
The `SearchEmptyState` template struct gained `has_filter: bool`,
`filter_empty_heading`, `filter_empty_body` fields. The fragment
template branches on `has_filter` (mirror of `home.html:407-428`).
Added `tracing::error!` log on render-error fallback.

P4 (blind, MEDIUM) — Restored outer-wrapper centering on home.html
empty branch + search_empty_state.html fragment. Both wrappers now
carry `class="text-center text-stone-500 dark:text-stone-400"` (with
`mt-12` on the SVG for top spacing) — fixes the visual regression
where the SVG sat flush-top because the macro's `py-12` was on its
inner div only. Avoids the doubled `py-12` AC6 originally accepted.

DEFERRED to GH Issues per Foundation Rule #11 (12 items):
- E2E parallel-mode flake risk on empty-states.spec.ts
- Locale cookie domain hardcoded `localhost`
- `<h2>` macro heading hierarchy a11y
- `data-variant` not validated against enum
- `search_empty_cta_url` computed unconditionally
- `body_html|safe` no caller-side enforcement (cross-cutting; already
  filed as GH #137 for modal + status_message sweep)
- Several tests with weak negative assertions
- `render_empty_state` admin role collapsed to librarian
- AC11/AC12 spec literal numeric drift (doc only)
- Macro role-gate values not validated
- i18n audit doesn't walk templates/
- `data-status-message` count not asserted in E2E

Post-patch validation:
- cargo check + cargo clippy --all-targets -- -D warnings — green
- cargo test --lib — 769 passed, 0 failed
- E2E single-spec empty-states.spec.ts + home-search.spec.ts — 20/20
  passed post `e2e-reset.sh` (incl. previously-flaky home-search:224)
- LOC: home.rs 1780 / home_search_fragment.rs 290 (both under 2000)
