+++ _bmad-output/implementation-artifacts/sprint-status.yaml
@@ -1,5 +1,5 @@
 # generated: 2026-03-29
-# last_updated: 2026-05-07 (Story 9-15 → review)
+# last_updated: 2026-05-08 (Story 9-15 → done)
 # project: mybibli
 # project_key: NOKEY
 # tracking_system: file-system
@@ -177,7 +177,7 @@ development_status:
   9-12-migrate-delete-contributor-to-modal: done
   9-13-migrate-delete-series-to-modal: done
   9-14-migrate-deactivate-user-to-modal: done
-  9-15-status-message-empty-states: review
+  9-15-status-message-empty-states: done
   9-16-connection-lost-overlay: backlog
   9-17-navbar-hamburger-and-scanner-autoclose: backlog
   9-18-navbar-role-visibility-polish: backlog
