+++ static/js/borrowers.js
@@ -11,16 +11,29 @@
         var form = document.getElementById("add-form");
         if (!form) return;
 
+        function showForm(e) {
+            if (e) e.preventDefault();
+            form.classList.remove("hidden");
+            var name = document.getElementById("new-name");
+            if (name) name.focus();
+        }
+
         var show = document.getElementById("borrowers-show-add-form");
         if (show && show.dataset.wired !== "true") {
             show.dataset.wired = "true";
-            show.addEventListener("click", function (e) {
-                e.preventDefault();
-                form.classList.remove("hidden");
-                var name = document.getElementById("new-name");
-                if (name) name.focus();
-            });
+            show.addEventListener("click", showForm);
         }
+
+        // Story 9-15 — the empty-state StatusMessage CTA also targets
+        // `#add-form`. Without this listener the CTA scrolls to the
+        // anchor but the form stays `display: none` (Tailwind `.hidden`),
+        // creating a UX dead-end on the encouraging empty state.
+        var emptyCtas = document.querySelectorAll('a[href="#add-form"]');
+        emptyCtas.forEach(function (cta) {
+            if (cta.dataset.wired === "true") return;
+            cta.dataset.wired = "true";
+            cta.addEventListener("click", showForm);
+        });
 
         var hide = document.getElementById("borrowers-hide-add-form");
         if (hide && hide.dataset.wired !== "true") {
