+++ templates/fragments/search_empty_state.html
@@ -1,9 +1,22 @@
 {# Story 9-15 — search empty-state fragment for live HTMX search.
    Mirrors the page-level home.html empty branch: SVG icon outside,
-   status_message macro inside (search variant with role-gated CTA). #}
+   status_message macro inside. Branches on `has_filter` to render
+   either the filter-empty (no CTA) or search-empty (librarian CTA)
+   sub-case — same contract as the navigated path. #}
 {% import "components/status_message.html" as status_message %}
-<div>
-    <svg class="mx-auto w-12 h-12 text-stone-300 dark:text-stone-600 mb-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"/></svg>
+<div class="text-center text-stone-500 dark:text-stone-400">
+    <svg class="mx-auto w-12 h-12 text-stone-300 dark:text-stone-600 mb-3 mt-12" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"/></svg>
+    {% if has_filter %}
+    {% call status_message::status_message(
+        "empty",
+        filter_empty_heading,
+        filter_empty_body,
+        "",
+        "",
+        "",
+        role,
+    ) %}{% endcall %}
+    {% else %}
     {% call status_message::status_message(
         "empty",
         search_empty_heading,
@@ -13,4 +26,5 @@
         "librarian",
         role,
     ) %}{% endcall %}
+    {% endif %}
 </div>
