+++ templates/pages/home.html
@@ -403,8 +403,8 @@
         <div id="browse-results" class="browse-list">
                 {% if let Some(paginated) = results %}
                     {% if paginated.items.is_empty() %}
-                    <div>
-                        <svg class="mx-auto w-12 h-12 text-stone-300 dark:text-stone-600 mb-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"/></svg>
+                    <div class="text-center text-stone-500 dark:text-stone-400">
+                        <svg class="mx-auto w-12 h-12 text-stone-300 dark:text-stone-600 mb-3 mt-12" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"/></svg>
                         {% if !active_filter.is_empty() %}
                         {% call status_message::status_message(
                             "empty",
